FIX: silence MSVC C5286 enum conversion warnings in ar_maxr/ar_minr
Cast cmpReals() return to int in switch expressions to avoid implicit
conversion from cmp_t to cmpex_t warnings on MSVC.

+++ src/pl-arith.c
@@ -3628,7 +3628,7 @@ ar_min(Number n1, Number n2, Number r)
 
 static bool
 ar_maxr(Number n1, Number n2, Number r)
-{ switch (cmpReals(n1, n2))
+{ switch ((int)cmpReals(n1, n2))
   { case CMP_LESS:
       cpNumber(r,n2);
       break;
@@ -3659,7 +3659,7 @@ ar_maxr(Number n1, Number n2, Number r)
 
 static bool
 ar_minr(Number n1, Number n2, Number r)
-{ switch (cmpReals(n1, n2))
+{ switch ((int)cmpReals(n1, n2))
   { case CMP_LESS:
       cpNumber(r,n1);
       break;
